move openPaymentChannel into core
